pyproject.toml
[project]
name = "image-captioning-mini"
version = "0.1.0"
description = ""
authors = [
    {name = "hauuto",email = "thanhhau670@gmail.com"}
]
package-mode = false
requires-python = ">=3.12"
dependencies = [
    "kagglehub (>=1.0.0,<2.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "nltk (>=3.9.4,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

